# server/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "2.0.0"
  title: HollaEx Kit
host: api.hollaex.com
basePath: /v1
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
  - text/csv
securityDefinitions:
  Bearer:
    description: JWT authentication
    type: apiKey
    name: Authorization
    in: header
  HmacKey:
    description: HMAC API key
    type: apiKey
    name: api-key
    in: header
paths:
...
  /user/request-otp:
    x-swagger-router-controller: otp
    get:
      operationId: requestOtp
      description: Generate OTP code for OTP verification
      tags:
        - User
        - OTP
      responses:
        200:
          description: Created
          schema:
            type: object
            required:
              - secret
            properties:
              secret:
                type: string
        403:
          description: Access Denied
          schema:
            $ref: "#/definitions/MessageResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/MessageResponse"
      security:
        - Bearer: []
      x-security-scopes:
        - user
  /user/activate-otp:
    x-swagger-router-controller: otp
    post:
      operationId: activateOtp
      description: Enable OTP
      parameters:
        - name: data
          description: Activation code
          in: body
          required: true
          schema:
            $ref: "#/definitions/OtpRequest"
      tags:
        - User
        - OTP
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/MessageResponse"
        403:
          description: Access Denied
          schema:
            $ref: "#/definitions/MessageResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/MessageResponse"
      security:
        - Bearer: []
      x-security-scopes:
        - user
  /user/deactivate-otp:
    x-swagger-router-controller: otp
    post:
      operationId: deactivateOtp
      description: Disable OTP
      tags:
        - User
        - OTP
      parameters:
        - name: data
          description: Activation code
          in: body
          required: true
          schema:
            $ref: "#/definitions/OtpRequest"
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/MessageResponse"
        403:
          description: Access Denied
          schema:
            $ref: "#/definitions/MessageResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/MessageResponse"
      security:
        - Bearer: []
      x-security-scopes:
        - user
definitions:
  MessageResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  OtpRequest:
    type: object
    required:
      - code
    properties:
      code:
        type: string